requirements.txt
aiocache==0.12.1
aiocqhttp==1.4.4
aiofiles==23.1.0
aiohttp==3.7.4.post0
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.7.0
APScheduler==3.10.4
async-timeout==3.0.1
attrs==23.1.0
blinker==1.5
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
distlib==0.3.6
et-xmlfile==1.1.0
filelock==3.10.7
fonttools==4.41.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
ifstools==1.15
itsdangerous==2.1.2
Jinja2==3.1.2
kbinxml==1.7
keyboard==0.13.5
kiwisolver==1.4.4
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.1
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pefile==2023.2.7
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.2.0
priority==2.0.0
psutil==5.9.5
pyparsing==3.0.9
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0.1
qq-botpy==1.1.2
qrcode==7.4.2
QtPy==2.3.1
quart==0.18.4
requests==2.30.0
scipy==1.11.1
seaborn==0.12.2
setuptools-scm==6.3.2
shiboken6==6.5.0
six==1.16.0
sniffio==1.3.0
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
typing_extensions==4.0.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
virtualenv==20.21.0
virtualenv-clone==0.5.7
websocket-client==1.6.2
Werkzeug==2.3.6
wsproto==1.2.0
yarl==1.9.2
zipp==3.6.0
mysql-connector-python==8.1.0
protobuf==4.21.12
